V1.0: Bank App & Server with Secure Connection
This release introduces version 1.0 of the Bank App and Server
application. The major enhancement is the implementation of Qssl
for secure connections. This ensures encrypted communication
between the app and the server.

Executables for this release are at the top level of the repository,
alongside the README. A key enhancement is the masking of password
fields. Now, they display asterisks instead of plaintext.

To run this application, generate your own private key and
certificate. Instructions for this are provided in the README.
For security reasons, a certificate or key is not provided.

This version is a fork of V0.5, which used an unsecured connection.
Future releases plan to enhance security by encrypting sensitive
data prior to database storage.